tests to make sure dependency helper functions are also called from the module we are testing
